fixed some bugs

+++ CAR_TRACKING/dt_GPU.cpp
@@ -598,8 +598,8 @@ CUT_THREADPROC dt_thread_func(void *p){
       }
 
       
-      pointer_dst_tmpIx += (unsigned long long int)(move_size * sizeof(FLOAT));
-      pointer_tmpIx_dev += (unsigned long long int)(move_size * sizeof(FLOAT));
+      pointer_dst_tmpIx += (unsigned long long int)(move_size * sizeof(int));
+      pointer_tmpIx_dev += (unsigned long long int)(move_size * sizeof(int));
 
       part_size = 0;
       pointer_size = 0;
@@ -695,8 +695,8 @@ CUT_THREADPROC dt_thread_func(void *p){
       }
 
       
-      pointer_dst_tmpIy += (unsigned long long int)(move_size * sizeof(FLOAT));
-      pointer_tmpIy_dev += (unsigned long long int)(move_size * sizeof(FLOAT));
+      pointer_dst_tmpIy += (unsigned long long int)(move_size * sizeof(int));
+      pointer_tmpIy_dev += (unsigned long long int)(move_size * sizeof(int));
 
       part_size = 0;
       pointer_size = 0;
